+++ backend/package.json
@@ -96,7 +96,7 @@
     "ssh2": "1.11.0",
     "stripe": "11.6.0",
     "swagger-ui-express": "4.6.0",
-    "tedious": "^15.1.2",
+    "tedious": "^15.1.3",
     "tunnel-ssh": "5.0.4",
     "typeorm": "^0.3.11",
     "uuid": "^9.0.0",
