+++ backend/test/ava-tests/table-postgres-e2e.test.ts
@@ -29,7 +29,7 @@ test.before(async () => {
     imports: [ApplicationModule, DatabaseModule],
     providers: [DatabaseService, TestUtils],
   }).compile();
-  app = moduleFixture.createNestApplication();
+  app = moduleFixture.createNestApplication() as any;
   testUtils = moduleFixture.get<TestUtils>(TestUtils);
   await testUtils.resetDb();
   app.use(cookieParser());
