+++ rocketadmin-agent/Dockerfile
@@ -1,4 +1,4 @@
-FROM node:16
+FROM node:18
 WORKDIR /app
 COPY package.json yarn.lock nest-cli.json /app/
 RUN yarn install --network-timeout 1000000
