+++ rocketadmin-agent/package.json
@@ -5,6 +5,7 @@
   "author": "",
   "private": true,
   "license": "UNLICENSED",
+  "type": "module",
   "scripts": {
     "prebuild": "rimraf dist",
     "build": "nest build",
@@ -14,26 +15,29 @@
     "start:debug": "nest start --debug --watch",
     "start:prod": "node dist/main",
     "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
-    "test": "jest",
+    "test": "jest --runInBand",
     "test:watch": "jest --watch",
     "test:cov": "jest --coverage",
     "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
     "test:e2e": "jest --config ./test/jest-e2e.json"
   },
   "dependencies": {
+    "@azure/core-tracing": "^1.0.1",
     "@nestjs/common": "^9.3.9",
     "@nestjs/config": "^2.3.1",
     "@nestjs/core": "^9.3.9",
     "@nestjs/platform-express": "^9.3.9",
     "@nestjs/platform-ws": "^9.3.9",
+    "@nestjs/testing": "^9.3.9",
+    "@types/pg": "^8.6.6",
     "argon2": "^0.30.3",
     "crypto-js": "^4.1.1",
     "faker": "^6.6.6",
     "get-port": "^6.1.2",
-    "knex": "^2.4.2",
-    "knex-paginate": "^3.0.2",
+    "knex": "2.4.2",
+    "knex-paginate": "3.0.2",
     "mysql2": "^3.1.2",
-    "nest-winston": "^1.8.0",
+    "nest-winston": "^1.9.0",
     "oracledb": "^5.5.0",
     "pg": "^8.9.0",
     "readline-sync": "^1.4.10",
