+++ rocketadmin-agent/src/dal/shared/create-dao.ts
@@ -1,10 +1,10 @@
-import { IDaoInterface } from './dao-interface';
-import { DaoPostgres } from '../dao/dao-postgres';
-import { DaoMysql } from '../dao/dao-mysql';
-import { DaoOracledb } from '../dao/dao-oracledb';
-import { DaoMssql } from '../dao/dao-mssql';
-import { ICLIConnectionCredentials } from '../../interfaces/interfaces';
-import { Messages } from '../../text/messages';
+import { IDaoInterface } from './dao-interface.js';
+import { DaoPostgres } from '../dao/dao-postgres.js';
+import { DaoMysql } from '../dao/dao-mysql.js';
+import { DaoOracledb } from '../dao/dao-oracledb.js';
+import { DaoMssql } from '../dao/dao-mssql.js';
+import { ICLIConnectionCredentials } from '../../interfaces/interfaces.js';
+import { Messages } from '../../text/messages.js';
 
 export function createDao(connection: ICLIConnectionCredentials): IDaoInterface {
   switch (connection.type) {
