+++ rocketadmin-agent/src/helpers/encryption/encryptor.ts
@@ -1,6 +1,6 @@
-import * as CryptoJS from 'crypto-js';
+import CryptoJS from 'crypto-js';
 import * as argon2 from 'argon2';
-import { Messages } from '../../text/messages';
+import { Messages } from '../../text/messages.js';
 
 export class Encryptor {
   public static encryptDataMasterPwd(data: string, masterPwd: string): string {
