+++ rocketadmin-agent/src/helpers/get-connection-to-db-params.ts
@@ -1,7 +1,7 @@
-import { ICLIConnectionCredentials } from '../interfaces/interfaces';
-import { toPrettyErrorsMsg } from './to-pretty-errors-msg';
-import { validateConnectionData } from './validate-connection-data';
-import { Config } from '../shared/config/config';
+import { ICLIConnectionCredentials } from '../interfaces/interfaces.js';
+import { toPrettyErrorsMsg } from './to-pretty-errors-msg.js';
+import { validateConnectionData } from './validate-connection-data.js';
+import { Config } from '../shared/config/config.js';
 
 export async function getConnectionToDbParams(): Promise<ICLIConnectionCredentials> {
   const connectionCredentials = Config.getConnectionConfig();
