+++ rocketadmin-agent/src/main.ts
@@ -1,17 +1,17 @@
-import * as WebSocket from 'ws';
+import WebSocket from 'ws';
 import { NestFactory } from '@nestjs/core';
-import { AppModule } from './app.module';
-import { CommandExecutor } from './command/command-executor';
-import { getConnectionToDbParams } from './helpers';
-import { OperationTypeEnum } from './enums/operation-type.enum';
-import { Messages } from './text/messages';
-import { checkConnection } from './helpers';
-import { ICLIConnectionCredentials } from './interfaces/interfaces';
-import { Config } from './shared/config/config';
-import { CLIQuestionUtility } from './helpers/cli/cli-questions';
-import { ConnectionTypeEnum } from './enums';
-import { mkDirIfNotExistsUtil } from './helpers/write-file-util';
-import { Constants } from './helpers/constants/constants';
+import { AppModule } from './app.module.js';
+import { CommandExecutor } from './command/command-executor.js';
+import { getConnectionToDbParams } from './helpers/get-connection-to-db-params.js';
+import { OperationTypeEnum } from './enums/operation-type.enum.js';
+import { Messages } from './text/messages.js';
+import { checkConnection } from './helpers/check-connection.js';
+import { ICLIConnectionCredentials } from './interfaces/interfaces.js';
+import { Config } from './shared/config/config.js';
+import { CLIQuestionUtility } from './helpers/cli/cli-questions.js';
+import { ConnectionTypeEnum } from './enums/connection-type.enum.js';
+import { mkDirIfNotExistsUtil } from './helpers/write-file-util.js';
+import { Constants } from './helpers/constants/constants.js';
 
 async function bootstrap() {
   const connectionCredentials: ICLIConnectionCredentials = Config.getConnectionConfig();
