+++ rocketadmin-agent/test/command/command-executor-mysql.spec.ts
@@ -1,14 +1,14 @@
-import * as faker from 'faker';
-import { AppModule } from '../../src/app.module';
+import { faker } from '@faker-js/faker';
+import { AppModule } from '../../src/app.module.js';
 import { INestApplication } from '@nestjs/common';
 import { Knex, knex } from 'knex';
 import { Test } from '@nestjs/testing';
-import { ICLIConnectionCredentials } from '../../src/interfaces/interfaces';
-import { CommandExecutor } from '../../src/command/command-executor';
-import { OperationTypeEnum } from '../../src/enums/operation-type.enum';
-import { Constants } from '../../src/helpers/constants/constants';
-import { QueryOrderingEnum } from '../../src/enums';
-import { DaoPostgres } from '../../src/dal/dao/dao-postgres';
+import { ICLIConnectionCredentials } from '../../src/interfaces/interfaces.js';
+import { CommandExecutor } from '../../src/command/command-executor.js';
+import { OperationTypeEnum } from '../../src/enums/operation-type.enum.js';
+import { Constants } from '../../src/helpers/constants/constants.js';
+import { QueryOrderingEnum } from '../../src/enums/query-ordering.enum.js';
+import { DaoPostgres } from '../../src/dal/dao/dao-postgres.js';
 
 describe('Command executor tests', () => {
   let app: INestApplication;
@@ -82,7 +82,7 @@ describe('Command executor tests', () => {
       imports: [AppModule],
     }).compile();
 
-    app = moduleFixture.createNestApplication();
+    app = moduleFixture.createNestApplication() as any;
     await app.init();
     await resetMySqlTestDB();
   });
@@ -632,7 +632,7 @@ describe('Command executor tests', () => {
         const commandExecutor = new CommandExecutor(connectionConfig);
 
         const settings = {
-          connection_id: faker.random.uuid(),
+          connection_id: faker.datatype.uuid(),
           table_name: testTableName,
           display_name: '',
           search_fields: [testTableColumnName],
@@ -679,7 +679,7 @@ describe('Command executor tests', () => {
           connection_id: faker.datatype.uuid(),
           table_name: testTableName,
           display_name: '',
-          search_fields: [faker.random.word(1)],
+          search_fields: [faker.random.words(1)],
           excluded_fields: undefined,
           list_fields: undefined,
           identification_fields: undefined,
