+++ rocketadmin-agent/tsconfig.json
@@ -1,15 +1,19 @@
 {
   "compilerOptions": {
-    "module": "commonjs",
+    "module": "ESNext",
     "declaration": true,
     "removeComments": true,
     "emitDecoratorMetadata": true,
     "experimentalDecorators": true,
-    "allowSyntheticDefaultImports": true,
-    "target": "es2017",
+    "allowJs": true,
+    "esModuleInterop": true,
+    "target": "ESNext",
     "sourceMap": true,
     "outDir": "./dist",
     "baseUrl": "./",
-    "incremental": true
-  }
+    "incremental": true,
+    "moduleResolution": "node"
+  },
+  "include": ["src", "test"],
+  "exclude": ["node_modules", "dist"]
 }
